+++ core/src/com/draga/Planet/Planet.java
@@ -9,7 +9,7 @@ import com.draga.GameEntity;
  */
 public class Planet extends GameEntity
 {
-    public Planet(float mass, int radius, int x, int y, String texturePath)
+    public Planet(float mass, float radius, float x, float y, String texturePath)
     {
         physicComponent = new CircularPhysicComponent(mass, radius, x, y);
         graphicComponent = new GraphicComponent(texturePath, physicComponent);
